backend/requirements.txt
pdfplumber==0.10.3
google-or-tools==9.8.3296
supabase-py==1.2.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
